`testing` module
Adds `bytewax.testing` as a module in the library which contains
helper tools for end users testing Bytewax dataflows.

The two things in there to start facilitate doctests:

1. `test_print()`: Have you ever tried printing dataflow output to
stdout but with multiple worker? Sometimes it is all interleaved and
that would make the output difficult to test. This wraps `print()` in
a mutex, so only one worker will print at a time. This allows for
deterministic output testing.

2. `doctest_ctx`: Doctests and spawning multiprocessing don't play
well together becuase there's no "main" Python file for the subprocess
to import to build back up state. This lets you specify a
`multiprocessing` "context" which doesn't actually spawn a pool of
subprocesses to do our local cluster work; instead everything happens
in process but in background threads. You can pass this to
`spawn_cluster()` or `run_cluster()`. Yes, this isn't "correct"
behavior, but the point of doctests is to demonstrate the API, and the
API and coarse behavior is unchanged. We still have proper pytests to
really test the true cluster behavior.