Updates longform docs to 0.8 and adds markdown tests
Adds an expected output block after every code block.

Many of the tests we turn out output sorting. We do this to ensure
that even if the workers returns epochs out of order, the test still
passes.

In some cases we turn out expected sorting. We do this to allow the
"out of order" answer to be in the docs to drive home the point that
the output here might not be ordered.

Lots of changes to get things working under 0.8. Mostly adding
`capture()` and using new stateful operator names.