Require state keys to be strings
I think it's going to be important to require the `key` in the `(key,
value)` 2-tuples to be a string. We / Timely do a lot with those keys
that we don't do with the values: shove them in hash maps (so they
need to properly be `Hash` and `PartialEq` and `Eq`), serialize them
and hope that the bytes are equal if the keys are equal (for recovery
DB lookup), print them for debugging. I'm worried that either folks
are going to use some Python types that don't obey equality correctly,
or there's going to be some issue with Pickling, and I think forcing
keys to be strings will make things easier to debug and write.

Adds a `StateKey` newtype wrapper around `String` and swaps out all
the `TdPyAny`s that represented the key with that type. Fills in some
shim traits.

Renames / cleans up `lift_2tuple` and `wrap_2tuple` into
`extract_state_pair` and `wrap_state_pair` with better error
reporting. Fixes https://github.com/bytewax/bytewax/issues/80

Tests the error messages and fixes docs to say the key has to be a
string.