Persist awake times
Prep
----

This turned into a sort of refresh of types used in the recovery
system to better match the model of what is actually happening.

The recovery store is modeled as two key-value DB tables, progress and
state, and the stateful operators generate CDC-style updates which are
applied to the tables. This PR renames the structs involved to match
that process. There's `ProgressUpdate` and `StateUpdate` which
correspond to KV changes. And each is a 2-tuple struct with 2 fields:
`ProgressRecoveryKey` and `ProgressOp` or `StateRecoveryKey` and
`StateOp`, with the `Op` enums representing what kind of change
happened to the value for each key (e.g. update or delete). Within
each `Op` is a `Progress` or `State` type that actually captures the
new value.

Stateful operators now emit fully formed `StateUpdate`s instead of the
difficult-to-name `EpochData`, which was the `(step_id, (state_key,
update))` tuple. This makes a nice symmetry between the backup phase
and the recovery loading, since everything uses the same types.

A lot of systemic renaming from `backup` to `update`, `update` to
`op`, and `state_bytes` to `state` to match this new convention.

Next Awake
----------

Now that we have a `State` type, we can add the `next_awake:
DateTime<Utc>` to that struct so we will save the next awake time for
each `StateKey`. Fills out that whenever we do a snapshot of the state
data in `StatefulUnary`.

Since all fields of `State` need to be written to each recovery store,
adds the `chrono` feature of `sqlx` to enable writing `DateTime` as a
row value. Updates the schema of the state table to add a `next_awake`
column.

Progress
--------

Makes both the SQLite and Kafka recovery stores actually store
progress data under the `ProgressRecoveryKey` to keep the parallel
construction between the state and progress recovery
tables. Previously, we were using the fixed string `"worker_frontier"`
which produces the same behavior when cluster size is static across
recovery, but this will set us up for our rescaling future where data
from multiple workers is mixed together. Instead the worker frontier
is one of the fields in the value (the only one as of now).

Adds a little `WorkerIndex` newtype to label when an int is meant to
be interpreted as the worker number.

As part of clarifying the datatypes needed to store progress, we don't
store the full antichain in the recovery store anymore; instead we
collapse the antichain into the minimum value during writing and only
store that single timestamp as the frontier. This helps with debugging
as you can now store the worker frontier as a single int, which is
much easier to parse out of the recovery store by hand rather than a
serde blob of bytes.

Reduces the logging level of recovery stores down to trace.