Store execution IDs and worker count for recovery
The progress recovery data model used to have an issue: we never knew
how many workers were part of a cluster. This caused us to "lose"
workers if we crashed before any progress data was written by a
worker; during resume, it just looked like the cluster was smaller and
nothing was missing. Adds a new kind of progress recovery message that
is written as each cluster starts up with the total worker count so we
can know if we're missing a later progress message.

Fixes https://github.com/bytewax/bytewax/issues/181

Related to this, we also need to disambiguate between workers across
clusters. During recovery we need to know which "worker 3" a progress
message "worker 3 moved to epoch 30" applies to. To do this we add an
`Execution` ID which is sort of like "epochs but for the whole
cluster". This is an `u64` that increments on each resume and all
progress recovery data is labeled with it so we ignore data from older
clusters that might come out of order. Adds tables to SQLite recovery
DB to do that.

This gets us one more step closer to rescaling.

Rewrites the `Progress` operator to work on frontiers again... Turns
out if a worker never generates input, it'll never generate state
data, which never generates progress info because it was based off of
the `Tick`s flowing through that part of the dataflow graph. This
means it's unnessecary gymnastics to store the `BorderEpoch` so
removes that from the progress data model.