Fix "interleaved executions"
Turns out they're ok. During resume, each worker's progress events are
in order. But! Because we're reading all the progress logs in parallel
and broadcasting them, across workers they might be out-of-order. Thus
we might do something like read all of worker 1's progress, then all
of worker 2's. Execution init messages are stored on all workers, so
the resume cluster might already have advanced to the final exeuction,
and then go "back in time" to see a slow reader's log. That's fine
because we already know the next exeuction started and thus can ignore
those progress messages.

Re-works some of the recovery tests to crash and restart twice to
start to probe this.