Revamp kafka input connector, schema registry, pyoperators
This commit revamps the whole Kafka input connector.
The connector is moved to its own module, split in different files.
The input connector now emits kafka messages with all the metadata included.
Support for Confluent's and Redpanda's schema registries was added, with
ability to (de)serialize messages to and from an avro schema.
The connector and the (de)serializations capabilities are also
exposed as a series of custom operators.