Passes a `now` argument to `build_part` and `next_batch`
This is a little backport of a little convenience from my operator
work. If you're using the `next_awake` feature of any of our inputs,
you need to know the current time when building and when the batch is
requested. It's really easy for us to pass that to you and makes it
less cumbersome to have to request the time in a few different spots.
I also think it helps guide people as to when you should update your
`next_awake` value.