Adds `get_fs_id` argument to `{Dir,File}Source`
There was a little bit of discussion of extending our file sources to
allow reading the same path, but support each worker process possibly
actually having different data at that path. Here's my take on that:

The `{Dir,File}Source` now takes an optional `get_fs_id` argument that
is called per directory and should return a consistent, unique ID for
the filesystem the directory is on. This `fs_id` is then added to the
input partition key so that we deduplicate on it.

This means if different workers have different data (assuming that the
same path on two different filesystems is different) at the same
paths, all unique data will be read; but if multiple workers mount the
same filesystem at the path, the partition assignment system ensures
you don't see duplicate data.

The tricky part here is finding the relevant `fs_id`. I had this
default to using `os.stat_result.st_dev` which seems to be a unique
int per filesystem. It is stable across restarts on my machine and
from what I can tell, should be consitent in a cluster if all machines
mount the same NFS filesystem. I do not think this will work with all
filesystems and FUSE, k8s volumes, etc will probably be different per
worker and accidentally result in duplicate data (because the `fs_id`
is different even though the underlying mounted filesystem is the
same). If someone is trying to use read from "local files" on a k8s
cluster, they'll need to do some testing and probably will override
this.

I do think this default handles the three most common use cases
correctly, though: 1 worker on a single machine (don't duplicate
data), N workers on a single machine (don't duplicate data), and N
workers on M machines (assume all data is unique).