Adds vermin pre-commit
Adds a pre-commit hook for https://github.com/netromdk/vermin which
checks that none of our Python code is using any features that don't
exist in 3.8, since that's the minimum version we want to support.

I keep accidentally using features only in 3.10 etc.

I couldn't find an easy way to have the command automatically
determine the minimum Python version from `pyproject.toml`, but
changing the one arg to keep in in sync with `requires-python` isn't
too big a deal.