Adds a simple retry functionality to `SimplePollingSource`
You can now raise a `SimplePollingSource.Retry(timeout: timedelta)` to
cause the `SimplePollingSource` to retry and play nice with
cooperative multitasking.

We can do this easily with the `SimplePollingSource` because we're
fully in control of the next awake logic. This exception would be very
edge-case-laden to add to input source in general because of the
extreme flexibility provided by the `*SourcePartition` ABC and needing
to think about when we should override the result of `next_awake` and
how to handle the state storage of that override. You just have to do
it yourself there to make all the relevant decisions.