Runtime typecheck `Stream` arguments to operators
If you don't have an LSP turned on, a common mistake is to forget to
"unpack" an operator that returns multiple streams. E.g. you
accidentally pass an entire `BranchOut` in where a `Stream` should go
instead of `BranchOut.trues`.

This usually manifests as a strange exception when the random type is
attempted to be scoped, and you get `BranchOut._scope` doesn't exist.
This isn't clear what's going on, so adds in some proactive type
checking when you call each operator so that each upstream is truly a
`Stream`.

This means we don't need to duplicate the "you forgot to unpack this"
message on every multiple return value.

Also gets rid of the `_AsArgs` protocol; we only used it for `Stream`
anyway and so made this more confusing then it needed to be.

Adds tests.