Changes snapshot and backup interval to saner defaults
Previously the defaults for this were backing up every 10 seconds and
keeping around snapshots for a day. This means your recovery DB would
max out at a size theoretically thousands of times bigger than your
in-memory state. This is "correct" but surprising.

Let's change the defaults to have only a 2x size increase. This is
probably a "somewhat resonable" default value for production but
should definitely be tuned for each deployment's backup configuration.