Removes `now` and `sched` arguments in input partitions and unary logic
I feel bad about this since it's breaking API churn, but I came back
upon this when I was doing the "bigger then memory state" prototyping.

I don't think we should include the `datetime` arguments to
`next_batch` and `on_item` in input partitions and stateful operators
for performance reasons.

From my benchmarking of stateful dataflows, a ton of the time is spent
converting these timestamps between Rust and Python, which is fine if
you use them, but it means that every input and stateful operator has
to pay the penalty even if they are not used.

By having to explicitly request the current time, you pay the penalty
if you need it, but all other operators are sped back up.

It also turns out that passing the scheduled times is redundant since
you'll always have that time stored in the partition / logic already
because you needed to return it from `next_awake` or `notify_at`, so
you should just access whatever instance var you returned there.