Substitute in the current Bytewax version into docs
Read the current defined version from `Cargo.toml`, save it as a
Sphinx substitution variable. Enable a Sphinx / MyST extension that
lets you put subsitutions in code blocks using `|var|` syntax. Use
that in all the spots in the docs.

Adds that extension as a dep into the reproducible venv.