Fixes for re-joining `WindowMetadata`
As part of thinking through
https://github.com/bytewax/bytewax/pull/466#discussion_r1607447828 and
how to properly re-join windowing operator output with metadata, I
realized there were a few little gotchas surrounding metadata and
joins that made actually performing that annoying.

Previously, the `windowing` operator (and thus all windowing operators
based on it) would emit metadata as soon as a window was opened, then
would emit updates whenever a window merged with it. This made it hard
to design the join afterwards, as you had no way of knowing for sure
that you had the final metadata. This PR now only emits the window
metadata once on window close. This means you know for sure you have
the final metadata and can focus on just thinking about how to
interpret the other side of the join.

It's now also possible to use the metadata as a "close trigger" in
some situations, since it is only emitted then.

`WindowMetadata` now includes a `merged_ids` set. This can be used to
re-construct more complicated session windowing situations. It's
possible that a windowing operator will emit data from a still open
window, so the data will be labeled with a window ID that later is
merged.

This change also adds some new modes to the `join` and `join_window`
operators to make these kinds of joins easier. There are now 2 mode
categories: "insert mode" which determines what happens when you get
multiple values on a side, and "emit mode" which determines when is a
join row emitted downstream. Having both these lets you do things like
"product complete" joins which would be useful for joining metadata to
windowing operators which emit multiple items downstream.

All of this makes doing the joins a little easier, but there is still
a lot of complexity in this area because the `windowing` operator is
so flexible.